requirements.txt
asgiref==3.7.2
Django
django-ckeditor==6.7.0
django-js-asset==2.1.0
sqlparse==0.4.4
tzdata==2023.3
gunicorn